Move snn stage to own file
