+++ snn_hfo_ieeg/stages/filter.py
@@ -1,7 +1,8 @@
 from typing import NamedTuple
 import numpy as np
-from snn_hfo_ieeg.functions.filter import *
-from snn_hfo_ieeg.functions.signal_to_spike import *
+from snn_hfo_ieeg.functions.filter import butter_bandpass_filter
+from snn_hfo_ieeg.functions.signal_to_spike import find_thresholds, signal_to_spike_refractory
+
 
 class Ripple(NamedTuple):
     up: np.array
